+++ src/app/main/todos/components/todo-item/todo-item.component.ts
@@ -1,5 +1,7 @@
 import { Component, Input } from '@angular/core';
-import { Todo } from '@shared/business/model/Todo';
+import { Store } from '@ngrx/store';
+import { Todo } from '@shared/business-domain/model/Todo';
+import { startUpdateTodoState } from '@todos/store/todolist/actions';
 
 @Component({
   selector: 'app-todo-item',
@@ -9,4 +11,14 @@ import { Todo } from '@shared/business/model/Todo';
 export class TodoItemComponent {
   @Input()
   public todo: Todo;
+
+  public get isDone(): boolean {
+    return Todo.isDone(this.todo);
+  }
+
+  public onChangeState(isDone: boolean): void {
+    this.store.dispatch(startUpdateTodoState({ todoToUpdate: this.todo, isDone }));
+  }
+
+  constructor(private store: Store) {}
 }