+++ src/app/main/todos/store/todolist/effects.ts
@@ -3,17 +3,37 @@ import { TodoHttpService } from '@todos/services/todo-http.service';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { of } from 'rxjs';
 import { catchError, map, mergeMap } from 'rxjs/operators';
-import { dataLoaded, loadingFailed, startLoading } from './actions';
+import {
+  listDataLoaded,
+  listLoadingFailed,
+  startLoadingList,
+  startUpdateTodoState,
+  todoStateUpdated,
+  todoStateUpdateFailed,
+} from './actions';
+import { TodoState } from '@shared/business-domain/model/TodoState';
 
 @Injectable()
 export class TodoListEffects {
   setTodoList$ = createEffect(() =>
     this.action$.pipe(
-      ofType(startLoading),
+      ofType(startLoadingList),
       mergeMap(() =>
         this.todoService.getTodos().pipe(
-          map((todos) => dataLoaded({ todos })),
-          catchError(() => of(loadingFailed()))
+          map((todos) => listDataLoaded({ todos })),
+          catchError(() => of(listLoadingFailed()))
+        )
+      )
+    )
+  );
+
+  updateTodo$ = createEffect(() =>
+    this.action$.pipe(
+      ofType(startUpdateTodoState),
+      mergeMap(({ todoToUpdate, isDone }) =>
+        this.todoService.updateTodo({ ...todoToUpdate, state: isDone ? TodoState.DONE : TodoState.TODO }).pipe(
+          map((updatedTodo) => todoStateUpdated({ updatedTodo })),
+          catchError(() => of(todoStateUpdateFailed()))
         )
       )
     )