+++ changelog.md
@@ -21,3 +21,15 @@
 - Adding aliases to make import readable in tsconfig.json
 - Adding an npm script to allow to run the application outside of virtual machine (port mapping)
 - Adding testing helper to make tests easier (observable, lazyloading)
+
+# [002] - Change a TODO state
+
+## Business
+
+- Allow user to update todo state ; if it became done, it will being placed as the last element
+
+## Technical
+
+- Renaming of todolist store actions
+- Renaming business folder to business-domain
+- Adding business rules for todos